requirements.txt
openai
python-dotenv
pypdf
openpyxl
faiss-cpu
numpy
tqdm
